+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Core01CpoOptimizationEngine.record_optimization_progress()
 - Core01DataManager APIs clear_optimization_progress/add_optimization_progress/get_optimization_progress_as_wide_df
 - core01_db_manager.Core01OptimizationProgressTable
+- Core01CpoOptimizationEngine.set_cpo_parameters now supports CPO `Workers` parameter based on `threads` parameter
 
 ## [0.5.7.1]- 2025-03-17
 ### Changed