+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]## [0.5.7.2b0]
+[Unreleased]## [0.5.7.2b1]
+
+## [0.5.7.2b0]
 ### Added
 - Optimization Progress Tracking in CPO. CPO callback to record progress of search. Supports reporting and visualization.
 - core01_cpo_optimization_engine.CpoProgressTrackerCallback